+++ enclave-manager/web/src/components/DownloadButton.tsx
@@ -1,15 +1,22 @@
-import { Button, ButtonProps } from "@chakra-ui/react";
+import { Button, ButtonProps, IconButton, IconButtonProps } from "@chakra-ui/react";
 import { FiDownload } from "react-icons/fi";
 import { isDefined } from "../utils";
 import { saveTextAsFile } from "../utils/download";
 
-type DownloadButtonProps = ButtonProps & {
+type DownloadButtonProps<IsIconButton extends boolean> = (IsIconButton extends true ? IconButtonProps : ButtonProps) & {
   valueToDownload?: (() => string) | string | null;
   fileName: string;
-  text?: string;
+  text?: IsIconButton extends true ? string : never;
+  isIconButton?: IsIconButton;
 };
 
-export const DownloadButton = ({ valueToDownload, text, fileName, ...buttonProps }: DownloadButtonProps) => {
+export const DownloadButton = <IsIconButton extends boolean>({
+  valueToDownload,
+  text,
+  fileName,
+  isIconButton,
+  ...buttonProps
+}: DownloadButtonProps<IsIconButton>) => {
   const handleDownloadClick = () => {
     if (isDefined(valueToDownload)) {
       const v = typeof valueToDownload === "string" ? valueToDownload : valueToDownload();
@@ -17,19 +24,32 @@ export const DownloadButton = ({ valueToDownload, text, fileName, ...buttonProps
     }
   };
 
-  if (!isDefined(valueToDownload)) {
+  if (!isDefined(valueToDownload) && !isDefined(buttonProps.onClick)) {
     return null;
   }
 
-  return (
-    <Button
-      leftIcon={<FiDownload />}
-      size={"xs"}
-      colorScheme={"darkBlue"}
-      onClick={handleDownloadClick}
-      {...buttonProps}
-    >
-      {text || "Download"}
-    </Button>
-  );
+  if (isIconButton) {
+    return (
+      <IconButton
+        icon={<FiDownload />}
+        size={"xs"}
+        variant={"ghost"}
+        colorScheme={"darkBlue"}
+        onClick={handleDownloadClick}
+        {...(buttonProps as IconButtonProps)}
+      />
+    );
+  } else {
+    return (
+      <Button
+        leftIcon={<FiDownload />}
+        size={"xs"}
+        colorScheme={"darkBlue"}
+        onClick={handleDownloadClick}
+        {...buttonProps}
+      >
+        {text || "Download"}
+      </Button>
+    );
+  }
 };